chore: Refactor the files and optimize the functionalities
